Added city_class.py file
The City class has the following attributes: name, country (where the city is located),
elevation (measured in meters), and population (approximate, according to recent
statistics). Fill in the blanks of the max_elevation_city function to return the
name of the city and its country (separated by a comma), when comparing the 3 defined
instances for a specified minimal population. For example, calling the function for a
minimum population of 1 million: max_elevation_city(1000000) should
return "Sofia, Bulgaria".