Creating new instances of a class object.
Creating new instances of class objects can be a great way to keep track of
values using attributes associated with the object. The values of these
attributes can be easily changed at the object level.
The following code illustrates a famous quote by George Bernard Shaw, using
objects to represent people. Fill in the blanks to make the code satisfy
the behavior described in the quote.