The function to check if a string is palindrome.
The is_palindrome function checks if a string is a palindrome.
A palindrome is a string that can be equally read from left to right or
right to left, omitting blank spaces, and ignoring capitalization.
Examples of palindromes are words like kayak and radar, and phrases
like "Never Odd or Even". Fill in the blanks in this function
to return True if the passed string is a palindrome, False if not.