Find the error in the code and fix it
The datetime module supplies classes for manipulating dates and times, and
contains many types, objects, and methods. You've seen some of them used in the
dow function, which returns the day of the week for a specific date.
We'll use them again in the next_date function, which takes the date_string parameter
in the format of "year-month-day", and uses the add_year function to calculate
the next year that this date will occur (it's 4 years later for the
29th of February during Leap Year, and 1 year later for all other dates).
Then it returns the value in the same format as it receives the date: "year-month-day".

Can you find the error in the code? Is it in the next_date function or the
add_year function? How can you determine if the add_year function returns
what it's supposed to? Add debug lines as necessary to find the problems,
then fix the code to work as indicated above.